requirements.txt
aiohttp==3.8.1
aioredis==1.3.1
beautifulsoup4==4.10.0
channels==3.0.4
channels-redis==3.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==4.0.2
django-filter==21.1
django-rest-framework==0.1.0
django-restframework==0.0.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
djoser==2.1.0
Jinja2==3.0.3
keyboard==0.13.5
PyJWT==2.3.0
pymorphy2==0.9.1
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0